refactor: Handle language from dev to prod

+++ packages/backend/src/core/admin/languages/services/delete.service.ts
@@ -1,20 +1,23 @@
 import { ABSOLUTE_PATHS } from '@/app.module';
 import { core_languages } from '@/database/schema/languages';
-import { configPath, ConfigType, getConfigFile } from '@/helpers/config';
 import { InternalDatabaseService } from '@/utils/database/internal_database.service';
 import {
   BadRequestException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { eq } from 'drizzle-orm';
 import { existsSync } from 'fs';
-import { rm, unlink, writeFile } from 'fs/promises';
+import { rm, unlink } from 'fs/promises';
 import { join } from 'path';
 
 @Injectable()
 export class DeleteLanguagesAdminService {
-  constructor(private readonly databaseService: InternalDatabaseService) {}
+  constructor(
+    private readonly databaseService: InternalDatabaseService,
+    private readonly configService: ConfigService,
+  ) {}
 
   async delete(id: number): Promise<void> {
     const language =
@@ -39,24 +42,29 @@ export class DeleteLanguagesAdminService {
       throw new BadRequestException('DEFAULT_LANGUAGE');
     }
 
-    const plugins = await this.databaseService.db.query.core_plugins.findMany({
-      orderBy: (table, { desc }) => desc(table.updated_at),
-      columns: {
-        code: true,
-      },
-    });
+    // Delete language files only in dev mode
+    if (this.configService.get('dev_mode')) {
+      const plugins = await this.databaseService.db.query.core_plugins.findMany(
+        {
+          orderBy: (table, { desc }) => desc(table.updated_at),
+          columns: {
+            code: true,
+          },
+        },
+      );
 
-    await Promise.all(
-      [...plugins, { code: 'core' }, { code: 'admin' }].map(async plugin => {
-        const path = join(
-          ABSOLUTE_PATHS.plugin({ code: plugin.code }).frontend.languages,
-          `${language.code}.json`,
-        );
+      await Promise.all(
+        [...plugins, { code: 'core' }, { code: 'admin' }].map(async plugin => {
+          const path = join(
+            ABSOLUTE_PATHS.plugin({ code: plugin.code }).frontend.languages,
+            `${language.code}.json`,
+          );
 
-        if (!existsSync(path)) return;
-        await unlink(path);
-      }),
-    );
+          if (!existsSync(path)) return;
+          await unlink(path);
+        }),
+      );
+    }
 
     // Remove assets
     const assetsPath = join(
@@ -65,11 +73,6 @@ export class DeleteLanguagesAdminService {
       language.code,
     );
     await rm(assetsPath, { recursive: true });
-
-    // Update config file
-    const config: ConfigType = getConfigFile();
-    config.langs = config.langs.filter(lang => lang.code !== language.code);
-    await writeFile(configPath, JSON.stringify(config, null, 2), 'utf-8');
 
     await this.databaseService.db
       .delete(core_languages)
